feat: remove product from cart

+++ hook-carrinho-compras/src/contexts/CartContext.tsx
@@ -1,7 +1,7 @@
 import { createContext, ReactNode, useContext, useState } from "react";
 import { toast } from "react-toastify";
 import { api } from "../services/api";
-import { Product } from "../types";
+import { Product, Stock } from "../types";
 
 interface CartProviderProps {
   children: ReactNode;
@@ -36,8 +36,21 @@ export const CartProvider = ({ children }: CartProviderProps) => {
 
   const addProduct = async (productId: number) => {
     try {
-      const newProduct = [...cart];
-      await api.get(`/stock/${productId}`);
+      const productAlreadyExists = cart.find(
+        (cartItem) => cartItem.id === productId
+      );
+
+      if (!productAlreadyExists) {
+        const { data: products } = await api.get<Product>(
+          `products/${productId}`
+        );
+
+        const { data: stock } = await api.get<Stock>(`stock/${productId}`);
+        if (stock.amount > 0) {
+        }
+        localStorage.setItem("@RocketShoes:cart", JSON.stringify(""));
+        return;
+      }
     } catch {
       toast.error("Quantidade solicitada fora de estoque");
     }
@@ -45,7 +58,16 @@ export const CartProvider = ({ children }: CartProviderProps) => {
 
   const removeProduct = (productId: number) => {
     try {
-      const newProduct = [...cart].filter((item) => item.id !== productId);
+      const productsExists = [...cart].some(
+        (cartItem) => cartItem.id === productId
+      );
+      if (!productsExists) {
+        toast.error("Erro na remoção do produto");
+        return;
+      }
+      const newProduct = [...cart].filter(
+        (cartItem) => cartItem.id !== productId
+      );
       setCart(newProduct);
       localStorage.setItem("@RocketShoes:cart", JSON.stringify(newProduct));
     } catch {
